new route

+++ api/routes/menuItems.js
@@ -6,7 +6,9 @@ const MenuItem = require('../models/menuItem');
 router.get('/', async (req, res, next) => {
     try {
         const result = await MenuItem.find()
-            .select("_id name")
+            .populate('menuIds')
+            .populate('typeArticle')
+            .populate('typeCategory')
             .exec();
         return res.status(200).json(result);
     } catch (err) {
@@ -15,10 +17,13 @@ router.get('/', async (req, res, next) => {
     }
 });
 
-router.get('/:categoryId', (req, res, next) => {
-    const id = req.params.categoryId;
+router.get('/:menuItemId', (req, res, next) => {
+    const id = req.params.menuItemId;
 
     MenuItem.findById(id)
+        .populate('menuIds')
+        .populate('typeArticle')
+        .populate('typeCategory')
         .exec()
         .then(doc => {
             console.log(doc);
@@ -35,13 +40,28 @@ router.get('/:categoryId', (req, res, next) => {
 });
 
 router.post('/insert', async (req, res, next) => {
-    const category = new MenuItem({
+
+    let data = {
         _id: new mongoose.Types.ObjectId,
         name: req.body.name,
-    });
+        menuIds: [],
+        datePosted: new Date(),
+    };
+
+    data.menuIds.push(req.body.menuIds);
+
+    if (req.body.typeArticle) {
+        data.typeArticle = req.body.typeArticle;
+    }
+
+    if (req.body.typeCategory) {
+        data.typeCategory = req.body.typeCategory;
+    }
+
+    const menuItem = new MenuItem(data);
 
     try {
-        const result = await category.save();
+        const result = await menuItem.save();
         return res.status(200).json(result);
     } catch (err) {
         console.log(err);
@@ -49,8 +69,8 @@ router.post('/insert', async (req, res, next) => {
     }
 });
 
-router.post('/update/:categoryId', async (req, res, next) => {
-    const id = req.params.categoryId;    
+router.post('/update/:menuItemId', async (req, res, next) => {
+    const id = req.params.menuItemId;
 
     try {
         const result = await MenuItem.updateMany({ _id: id }, { $set: req.body }).exec()
@@ -65,13 +85,16 @@ router.post('/update/:categoryId', async (req, res, next) => {
 
 router.post('/delete', async (req, res, next) => {
 
-    const categoryIds = req.body.categoryIds;
-
     try {
+        if (!req.body.menuItemIds) {
+            throw new Error("Please provide appropriate parameters");
+        }
+
         const result = await MenuItem.deleteMany({
-            _id: {$in: categoryIds}
+            _id: { $in: req.body.menuItemIds }
         }).exec();
-        res.status(200).json(result);
+
+        return res.status(200).json(result);
     } catch (err) {
         console.log(err);
         return res.status(500).json({
